clas-algo-openbci (2)/requirements.txt
matplotlib
pandas
numpy
scipy
simpleaudio
colorednoise
fooof
pyserial
pyyaml
pyqt5
qdarkstyle
psutil
requests
